Add ability to export scans as HTML
This allows a single HTML page to be downloaded with a static version
of the distiller app along with the embedded scan data. To do this
we build a single page version of the app with various features disabled.
This is used as a template to insert the appriopriate scan data using
handlebars.js when a user chooses to export scans as HTML.